+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.2.0
+* Fix: Resolve `Image` block import issues.
+* Fix: Missing post title during import.
+* Feat: Add custom hooks `cbtj_import_block`, `cbtj_export_block`, `cbtj_blocks`.
+* Test: Add PHP unit tests to improve code.
+* Docs: Update README docs.
+* Tested up to WP 6.8.
+
 ## 1.1.0
 * Feat: Add REST namespace filter `cbtj_rest_namespace`.
 * Refactor: Use classes for PHP codebase.